#18 feat : ocr 기능 분리
func.py에 기능 정의. main에서 불러올 때 인자 값으로 img_path 변수 생성해서 넣어줘야 run_ocr의 self.img_path에 전달된다. (= 동일한 변수명으로 전달 시 자동으로 바인딩된다)
