--- ocr.py
@@ -1,24 +1,0 @@
-from pororo import Pororo
-import time
-
-class PororoOcr:
-    def __init__(self, model: str = "brainocr", lang: str = "ko", **kwargs):
-        self.model = model
-        self.lang = lang
-        self._ocr = Pororo(task="ocr", lang=lang, model=model, **kwargs)
-        self.img_path = None
-        self.ocr_result = {}
-
-    def run_ocr(self, img_path: str, debug: bool = False):
-        self.img_path = img_path
-        self.ocr_result = self._ocr(img_path, detail=True)
-
-        if self.ocr_result['description']:
-            ocr_text = self.ocr_result["description"]
-        else:
-            ocr_text = "No text detected."
-
-        if debug:
-            self.show_img_with_ocr()
-
-        return ocr_text
